application/requirements.txt
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
pandas==2.0.3
psycopg2==2.9.7  # For PostgreSQL
